--- GroundStation/test.py
@@ -1,5 +1,0 @@
-import time
-print('text one', end = '\r')
-time.sleep(1)
-print('text two', end = '\r')
-time.sleep(2)
